Add look up/crouching down turntable sprites
This adds in custom sprites and accompanying code for looking up and crouching down for the turntable animation; i.e. CNZ barrels, HCZ spinning pillars. Note that the sprites for these are custom and are based off the original S3K sprites, not Origins. Tails also does not have any sprites for this currently.